+++ .github/workflows/pr.yml
@@ -17,12 +17,15 @@ jobs:
       # Clone the logos-delivery checkout OUTSIDE the module tree: a directory
       # named `vendor/` at the module root would put Go into vendor mode.
       LOGOS_DELIVERY_DIR: ${{ github.workspace }}/.logos-delivery
-      # Both kernel (libwaku) and messaging (liblogosdelivery) headers are on
-      # the include path so every package compiles. `-l` is deliberately NOT
-      # set here: each internal/ffi subpackage selects its own library via a
-      # `#cgo LDFLAGS` directive, so no binary links both libs (they export
-      # overlapping symbols until logos-delivery#3851).
-      CGO_CFLAGS: -I${{ github.workspace }}/.logos-delivery/library/ -I${{ github.workspace }}/.logos-delivery/liblogosdelivery/
+      # TODO(revert once merged): pinned to logos-delivery#4012, which unifies
+      # the node lifecycle on logosdelivery_* and ships the full API in the
+      # single liblogosdelivery library. Point back at the default branch after
+      # it merges.
+      LOGOS_DELIVERY_REF: update-examples-ffi-node-lifecycle
+      # The single liblogosdelivery library exposes the full API; its headers
+      # live in library/. The bridge self-links -llogosdelivery via a #cgo
+      # directive, so no -l is set here.
+      CGO_CFLAGS: -I${{ github.workspace }}/.logos-delivery/library/
       CGO_LDFLAGS: -L${{ github.workspace }}/.logos-delivery/build/ -Wl,-rpath,${{ github.workspace }}/.logos-delivery/build/
       # Build in module mode; never use a vendor/ dir.
       GOFLAGS: -mod=mod
@@ -39,12 +42,11 @@ jobs:
           go-version: "1.24"
 
       - name: Resolve logos-delivery commit
-        # Cache the built libraries keyed on the exact upstream commit, so the
-        # expensive clone + build is skipped while logos-delivery's HEAD is
-        # unchanged. ls-remote gives us the SHA before we clone.
+        # Cache the built library keyed on the exact upstream commit, so the
+        # expensive clone + build is skipped while the pinned ref is unchanged.
         id: logos-delivery-rev
         run: |
-          rev=$(git ls-remote https://github.com/logos-messaging/logos-delivery.git HEAD | cut -f1)
+          rev=$(git ls-remote https://github.com/logos-messaging/logos-delivery.git "$LOGOS_DELIVERY_REF" | cut -f1)
           echo "rev=$rev" >> "$GITHUB_OUTPUT"
 
       - name: Cache logos-delivery build
@@ -56,11 +58,11 @@ jobs:
 
       - name: Clone logos-delivery
         if: steps.logos-delivery-cache.outputs.cache-hit != 'true'
-        run: git clone --depth 1 https://github.com/logos-messaging/logos-delivery.git "$LOGOS_DELIVERY_DIR"
+        run: git clone --depth 1 --branch "$LOGOS_DELIVERY_REF" https://github.com/logos-messaging/logos-delivery.git "$LOGOS_DELIVERY_DIR"
 
-      - name: Build libwaku + liblogosdelivery
+      - name: Build liblogosdelivery
         if: steps.logos-delivery-cache.outputs.cache-hit != 'true'
-        run: make -C "$LOGOS_DELIVERY_DIR" libwaku liblogosdelivery -j
+        run: make -C "$LOGOS_DELIVERY_DIR" liblogosdelivery -j
 
       - name: go build
         run: go build ./...
@@ -88,9 +90,9 @@ jobs:
 
       - name: go test messaging (run)
         # Fast, network-free unit tests for the Messaging API + its ffi bridge.
-        run: go test ./pkg/messaging/... ./internal/ffi/liblogosdelivery/...
+        run: go test ./pkg/messaging/... ./internal/ffi/...
 
       - name: go test kernel (compile)
         # The kernel suite is heavy integration (runs nightly in CI.yml), so
         # only compile its test binaries here.
-        run: go test -run '^$' ./pkg/kernel/... ./internal/ffi/libwaku/...
+        run: go test -run '^$' ./pkg/kernel/...
