+++ pkg/kernel/doc.go
@@ -1,3 +1,4 @@
 // Package kernel is the low-level Go wrapper over the logos-delivery Kernel API
-// (libwaku): relay, store, lightpush, filter, peer management, discovery.
+// (the waku_* tier of the single liblogosdelivery library): relay, store,
+// lightpush, filter, peer management, discovery.
 package kernel
